moves test classes to new folders reflected by main lib
