+++ tests/Pliant.Tests.Unit/Runtime/AycockHorspoolAlgorithmTests.cs
@@ -5,7 +5,7 @@ using Pliant.Grammars;
 using Pliant.Runtime;
 using Pliant.Tokens;
 
-namespace Pliant.Tests.Unit
+namespace Pliant.Tests.Unit.Runtime
 {
     [TestClass]
     public class AycockHorspoolAlgorithmTests
