+++ tests/Pliant.Tests.Unit/Runtime/ParseEngineTests.cs
@@ -9,7 +9,7 @@ using Pliant.Tests.Unit.Forest;
 using Pliant.Builders.Expressions;
 using Pliant.Runtime;
 
-namespace Pliant.Tests.Unit
+namespace Pliant.Tests.Unit.Runtime
 {
     [TestClass]
     public class ParseEngineTests
