+++ tests/Pliant.Tests.Unit/Runtime/ParseRunnerTests.cs
@@ -6,7 +6,7 @@ using Pliant.Runtime;
 using Pliant.Tokens;
 using System;
 
-namespace Pliant.Tests.Unit
+namespace Pliant.Tests.Unit.Runtime
 {
     [TestClass]
     public class LexerTests
